+++ src/UI/SignTransactionsModals/components/confirmationScreen.types.ts
@@ -1,0 +1,7 @@
+import { SignPropsType } from '../types/signTransactionsModals.types';
+
+export interface ExtraConfirmationScreenPropsType {
+  Screen?: (signProps: SignPropsType) => JSX.Element;
+  verifyReceiverScam?: boolean;
+  className?: string;
+}