+++ src/hooks/transactions/index.ts
@@ -1,6 +1,7 @@
 export * from './useGetTransactionDisplayInfo';
 export * from './useParseSignedTransactions';
 export * from './useSignTransactions';
+export * from './useSignTransactionsCommonData';
 export * from './useGetSignedTransactions';
 export * from './useGetTokenDetails';
 export * from './useSignTransactionsWithDevice';