+++ CHANGELOG.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[2.1.11](https://github.com/ElrondNetwork/dapp-core/pull/474)] - 2022-11-15
+- [Changed `ConfirmationScreen` calling sign hooks depeding on device](https://github.com/ElrondNetwork/dapp-core/pull/461)
+- [Updated legacy peers type definitions](https://github.com/ElrondNetwork/dapp-core/pull/460)
+
 ## [[2.1.10](https://github.com/ElrondNetwork/dapp-core/pull/473)] - 2022-11-11
 - [Fixed `getTrimmedHash` helper](https://github.com/ElrondNetwork/dapp-core/pull/473)
+
 ## [[2.1.9](https://github.com/ElrondNetwork/dapp-core/pull/472)] - 2022-11-11
 - [Added `getTransactionLinkWithLabel` helper](https://github.com/ElrondNetwork/dapp-core/pull/471)
 
@@ -21,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [[2.1.6](https://github.com/ElrondNetwork/dapp-core/pull/463)] - 2022-11-03
 - [Removed `signTransactionsCancelMessage` from persisted state](https://github.com/ElrondNetwork/dapp-core/pull/462)
+- [Changed `useSignTransactions` being called for ledger and extra provider](https://github.com/ElrondNetwork/dapp-core/pull/461)
 
 ## [[2.1.5](https://github.com/ElrondNetwork/dapp-core/pull/459)] - 2022-11-02
 - [Changed `useSignTransaction` hook to prevent multiple sign requests on same transactions](https://github.com/ElrondNetwork/dapp-core/pull/458)